Minor fixes to data type and conf_max

+++ uarchsim/payload.h
@@ -199,10 +199,10 @@ typedef struct {
 
    unsigned int lane_id; // Execution lane chosen for the instruction.
 
-   uint64_t Predicted_value;  // Contains value predicted by SVP, if eligibil and confident
+   int64_t Predicted_value;  // Contains value predicted by SVP, if eligibil and confident
    bool     vp_eligible; // tells whether the value is vp eligible or not
    int vp_conf;
-   int vpq_index;
+   uint64_t vpq_index;
    ////////////////////////
    // Set by Issue Stage.
    ////////////////////////
